Fix for 0.00 values.

+++ server/sql/db_1.sql
@@ -11,8 +11,8 @@ CREATE TABLE IF NOT EXISTS `version` (
 CREATE TABLE IF NOT EXISTS `characters` (
   `identifier` varchar(50) COLLATE utf8mb4_bin DEFAULT NULL,
   `characterid` int(11) DEFAULT 0,
-  `money` int(11) DEFAULT 0,
-  `gold` int(11) DEFAULT 0,
+  `money` double(11, 2) DEFAULT 0,
+  `gold` double(11, 2) DEFAULT 0,
   `xp` int(11) DEFAULT 0,
   `level` int(11) DEFAULT 0,
   `job` varchar(50) COLLATE utf8mb4_bin DEFAULT 'unemployed',
@@ -25,8 +25,8 @@ CREATE TABLE IF NOT EXISTS `users` (
   `identifier` varchar(50) COLLATE utf8mb4_bin NOT NULL,
   `license` varchar(50) COLLATE utf8mb4_bin DEFAULT NULL,
   `name` varchar(255) CHARACTER SET utf8 COLLATE utf8_bin DEFAULT NULL,
-  `money` int(11) DEFAULT NULL,
-  `bank` int(11) DEFAULT NULL,
+  `money` double(11, 2) DEFAULT NULL,
+  `bank` double(11, 2) DEFAULT NULL,
   `permission_level` int(11) DEFAULT NULL,
   `group` varchar(50) COLLATE utf8mb4_bin DEFAULT NULL,
   PRIMARY KEY (`identifier`)
